refactor: move service and run-server out of lib root.
* split into bin/ and the new http/ which will both soon contain more
  things and better structure.
