+++ lib/bin/run-server.js
@@ -5,11 +5,11 @@ const exit = require('express-graceful-exit');
 // CONTAINER SETUP
 
 // initialize our top-level static database instance.
-const { connect } = require('./model/database');
+const { connect } = require('../model/database');
 const db = connect();
 
 // set up our mailer.
-const { mailer } = require('./outbound/mail');
+const { mailer } = require('../outbound/mail');
 const mail = mailer(config.get('default.email'));
 
 
@@ -17,8 +17,8 @@ const mail = mailer(config.get('default.email'));
 // START HTTP SERVICE
 
 // initialize our container, then generate an http service out of it.
-const container = require('./model/package').withDefaults({ db, mail });
-const service = require('./service')(container);
+const container = require('../model/package').withDefaults({ db, mail });
+const service = require('../http/service')(container);
 
 // insert the graceful exit middleware.
 service.use(exit.middleware(service));
