+++ lib/http/service.js
@@ -18,7 +18,7 @@ module.exports = (container) => {
 
   // pull session and version information and pass it on, or reject the request if
   // broken information is provided for either.
-  const { versionParser, sessionParser, fieldKeyParser, headerOptionsParser } = require('./util/http');
+  const { versionParser, sessionParser, fieldKeyParser, headerOptionsParser } = require('../util/http');
   service.use(versionParser);
   service.use(sessionParser(container));
   service.use(fieldKeyParser(container));
@@ -28,19 +28,19 @@ module.exports = (container) => {
   ////////////////////////////////////////////////////////////////////////////////
   // RESOURCES
 
-  require('./resources/field-keys')(service, container);
-  require('./resources/forms')(service, container);
-  require('./resources/users')(service, container);
-  require('./resources/sessions')(service, container);
-  require('./resources/submissions')(service, container);
-  require('./resources/config')(service, container);
+  require('../resources/field-keys')(service, container);
+  require('../resources/forms')(service, container);
+  require('../resources/users')(service, container);
+  require('../resources/sessions')(service, container);
+  require('../resources/submissions')(service, container);
+  require('../resources/config')(service, container);
 
 
   ////////////////////////////////////////////////////////////////////////////////
   // POSTRESOURCE HANDLERS
 
   // apply output error handler to everything.
-  const { sendError } = require('./util/http');
+  const { sendError } = require('../util/http');
   service.use((error, request, response, next) => {
     if (response.headersSent === true) {
       // In this case, we'll just let Express fail the request out.
